requirements.txt
pyinstaller==5.13.2
pillow==9.5.0
